Refactored computeAreaIntegral pulling internals out into a computeAreaIntegralInternal function that is vectorized.
In the case that farea is not true, then callable(clm) is called, which
is not vectorizable. Computation to that point can be stored as dF and
the Jacobian. These are of size NSTxNPxNPx3 for dF and NSTxNPxNP for the
Jacobian. Storing these (only in the case of farea being false), the
internals can still be vectorized with a follow on loop added to
calculate dFunIntegral.